Added CitedbyInfo n updated ingestion directory
