+++ server/app/routers/elastic_routes.py
@@ -4,23 +4,21 @@ from uuid import uuid4
 from fastapi import APIRouter
 
 from models.api_models import SearchQueryResponse, PaperDetailResponse, CitationsResponse, ClusterDetailResponse, \
-    showCitingClustersResponse, SearchQuery
-from models.domain_models import Paper, Citation, Cluster
-from services.citation_adapter import CitationDao
-from services.cluster_adapter import ClusterDao
+    showCitingClustersResponse, SearchQuery, Paper, Citation, Cluster
+
 from services.elastic_service import ElasticService
-from services.paper_adapter import PaperDao
+from services.elasticsearch_adapters import PaperAdapter, CitationAdapter, ClusterAdapter
 
 router = APIRouter()
 elastic_service = ElasticService()
-paper_dao = PaperDao(elastic_service)
-cluster_dao = ClusterDao(elastic_service)
-citation_dao = CitationDao(elastic_service)
+paper_adapter = PaperAdapter(elastic_service)
+cluster_adapter = ClusterAdapter(elastic_service)
+citation_adapter = CitationAdapter(elastic_service)
 
 
 @router.post('/search', response_model=SearchQueryResponse)
 def perform_search(searchQuery: SearchQuery):
-    docs_response = paper_dao.search_papers(searchQuery)
+    docs_response = paper_adapter.search_papers(searchQuery)
     result_list = []
     for doc_hit in docs_response['hits']['hits']:
         result_list.append(build_paper_entity(doc=doc_hit['_source']))
@@ -30,14 +28,14 @@ def perform_search(searchQuery: SearchQuery):
 
 @router.get('/paper/{id}')
 def paper_info(id: str):
-    docs_response = paper_dao.get_paper_info(id)
+    docs_response = paper_adapter.get_paper_info(id)
     paper_entity_response = build_paper_entity(docs_response['hits']['hits'][0]['_source'])
     return PaperDetailResponse(query_id=str(uuid4()), paper=paper_entity_response)
 
 
 @router.get('/citations/{id}')
 def citations(id: str, page: int = 1, pageSize: int = 10):
-    es_citations_response = citation_dao.get_citations_for_paper(id, page, pageSize)
+    es_citations_response = citation_adapter.get_citations_for_paper(id, page, pageSize)
     result_list = []
     for doc_hit in es_citations_response['hits']['hits']:
         result_list.append(build_citation_entity(doc=doc_hit['_source']))
@@ -47,22 +45,22 @@ def citations(id: str, page: int = 1, pageSize: int = 10):
 
 @router.get('/cluster/{cid}')
 def show_cluster_detail(cid: str):
-    cluster_detail_response = cluster_dao.get_cluster_info(cid)
+    cluster_detail_response = cluster_adapter.get_cluster_info(cid)
     return ClusterDetailResponse(query_id=str(uuid4()),
                                  cluster=build_cluster_entity(cluster_detail_response['_source']))
 
 
 @router.get('/showCiting/{cid}')
 def show_citing(cid: str, sort: str, page: int, pageSize: int):
-    cluster_response = cluster_dao.get_cluster_info(cid)
+    cluster_response = cluster_adapter.get_cluster_info(cid)
     primary_cluster_detail = build_cluster_entity(cluster_response['_source'])
     cid_list = cluster_response['_source']['cited_by']
-    response = cluster_dao.get_paper_ids_for_clusters(cid_list=cid_list)
+    response = cluster_adapter.get_paper_ids_for_clusters(cid_list=cid_list)
     papers_list = []
     for each_hit in response['docs']:
         if 'included_papers' in each_hit['_source'] and len(each_hit['_source']['included_papers']) != 0:
             papers_list.append(each_hit['_source']['included_papers'][0])
-    papers_response = paper_dao.get_sorted_papers(papers_list=papers_list, page=page, pageSize=pageSize, sort=sort)
+    papers_response = paper_adapter.get_sorted_papers(papers_list=papers_list, page=page, pageSize=pageSize, sort=sort)
     result_list = []
     for each_paper_hit in papers_response['hits']['hits']:
         result_list.append(build_paper_entity(doc=each_paper_hit['_source']))
@@ -73,7 +71,7 @@ def show_citing(cid: str, sort: str, page: int, pageSize: int):
 
 @router.get('/similar')
 def similar_papers(paperID: str = ""):
-    result = cluster_dao.get_clustered_papers(paperID)
+    result = cluster_adapter.get_clustered_papers(paperID)
     return result
 
 
@@ -82,7 +80,8 @@ def build_paper_entity(doc):
                  title=doc.get('title'),
                  venue=doc.get('venue'),
                  year=doc.get('year'),
-                 n_citation=doc.get('ncites'),
+                 n_cited_by=doc.get('ncites'),
+                 n_self_cites=doc.get('selfCites'),
                  abstract=doc.get('abstract'),
                  bibtex="test_bibtex",
                  authors=get_authors_in_list(doc, 'authors'),
